This commit was manufactured by cvs2git to create tag 'glite-lb_R_1_10_0_2'.
Sprout from branch_RC31_3 2010-08-19 14:30:10 UTC Zdeněk Šustr <sustr4@cesnet.cz> 'New age for the subsys'
Cherrypick from master 2010-06-25 09:08:51 UTC Michal Voců <michal@ruk.cuni.cz> 'plugin framework fixes':
    org.glite.lb.harvester/doc/INSTALL
    org.glite.lb.harvester/doc/README
    org.glite.lb.harvester/doc/glite-lb-harvester.sgml
    org.glite.lb.harvester/examples/test.sql
    org.glite.lb.harvester/project/package.summary
    org.glite.lb.logger/Makefile
    org.glite.lb.logger/config/startup
    org.glite.lb.logger/configure
    org.glite.lb.logger/doc/glite-lb-interlogd.8
    org.glite.lb.logger/doc/glite-lb-logd.8
    org.glite.lb.logger/interface/il_error.h
    org.glite.lb.logger/interface/interlogd.h
    org.glite.lb.logger/project/ChangeLog
    org.glite.lb.logger/project/package.description
    org.glite.lb.logger/project/package.summary
    org.glite.lb.logger/project/version.properties
    org.glite.lb.logger/src-nt/Connection.H
    org.glite.lb.logger/src-nt/Connection.cpp
    org.glite.lb.logger/src-nt/EventManager.H
    org.glite.lb.logger/src-nt/EventManager.cpp
    org.glite.lb.logger/src-nt/Exception.H
    org.glite.lb.logger/src-nt/HTTPTransport.H
    org.glite.lb.logger/src-nt/HTTPTransport.cpp
    org.glite.lb.logger/src-nt/InputChannel.H
    org.glite.lb.logger/src-nt/InputChannel.cpp
    org.glite.lb.logger/src-nt/Makefile
    org.glite.lb.logger/src-nt/Message.H
    org.glite.lb.logger/src-nt/MessageStore.H
    org.glite.lb.logger/src-nt/MessageStore.cpp
    org.glite.lb.logger/src-nt/PlainConnection.H
    org.glite.lb.logger/src-nt/PlainConnection.cpp
    org.glite.lb.logger/src-nt/PluginManager.H
    org.glite.lb.logger/src-nt/PluginManager.cpp
    org.glite.lb.logger/src-nt/Properties.H
    org.glite.lb.logger/src-nt/Singleton.H
    org.glite.lb.logger/src-nt/SocketInput.H
    org.glite.lb.logger/src-nt/SocketInput.cpp
    org.glite.lb.logger/src-nt/ThreadPool.H
    org.glite.lb.logger/src-nt/ThreadPool.cpp
    org.glite.lb.logger/src-nt/Transport.H
    org.glite.lb.logger/src-nt/Transport.cpp
    org.glite.lb.logger/src-nt/main.cpp
    org.glite.lb.logger/src-nt/test/EventManagerTest.cpp
    org.glite.lb.logger/src-nt/test/PluginManagerTest.cpp
    org.glite.lb.logger/src-nt/test/SingletonTest.cpp
    org.glite.lb.logger/src-nt/test/ThreadPoolTest.cpp
    org.glite.lb.logger/src-nt/test/test_main.cpp
    org.glite.lb.logger/src/activemq_cpp_plugin.cpp
    org.glite.lb.logger/src/event_queue.c
    org.glite.lb.logger/src/event_store.c
    org.glite.lb.logger/src/event_store_http.c
    org.glite.lb.logger/src/http.c
    org.glite.lb.logger/src/il_error.c
    org.glite.lb.logger/src/il_master.c
    org.glite.lb.logger/src/input_queue_socket.c
    org.glite.lb.logger/src/input_queue_socket_http.c
    org.glite.lb.logger/src/interlogd.c
    org.glite.lb.logger/src/logd.c
    org.glite.lb.logger/src/logd_proto.c
    org.glite.lb.logger/src/logd_proto.h
    org.glite.lb.logger/src/perftest_il.sh
    org.glite.lb.logger/src/perftest_ll.sh
    org.glite.lb.logger/src/plugin_mgr.c
    org.glite.lb.logger/src/queue_mgr.c
    org.glite.lb.logger/src/queue_mgr_http.c
    org.glite.lb.logger/src/queue_thread.c
    org.glite.lb.logger/src/recover.c
    org.glite.lb.logger/src/send_event.c
    org.glite.lb.logger/src/send_event_http.c
    org.glite.lb.logger/src/server_msg.c
    org.glite.lb.logger/src/server_msg_http.c
    org.glite.lb.logger/test/IlTestBase.cpp
    org.glite.lb.logger/test/IlTestBase.h
    org.glite.lb.logger/test/event_queueTest.cpp
    org.glite.lb.logger/test/event_storeTest.cpp
    org.glite.lb.logger/test/il_test.cpp
    org.glite.lb.logger/test/input_queue_socketTest.cpp
    org.glite.lb.logger/test/ll_test.cpp
    org.glite.lb.logger/test/logd_proto_test.c
    org.glite.lb.logger/test/server_msgTest.cpp
Cherrypick from master 2010-03-10 12:17:51 UTC Zdeněk Šustr <sustr4@cesnet.cz> 'The most recent version copied. Do not modify this instance (RW in ./org.glite.lb).':
    org.glite.lb.harvester/Makefile
    org.glite.lb.harvester/configure
    org.glite.lb.harvester/examples/test.sh
    org.glite.lb.harvester/project/ChangeLog
    org.glite.lb.harvester/project/package.description
    org.glite.lb.harvester/project/version.properties
    org.glite.lb.harvester/src/harvester.c
Delete:
    org.glite.lb.logger/src/il_error.h
    org.glite.lb.logger/src/interlogd.h

+++ org.glite.lb.logger/src-nt/SocketInput.cpp
@@ -1,3 +1,20 @@
+/*
+Copyright (c) Members of the EGEE Collaboration. 2004-2010.
+See http://www.eu-egee.org/partners for details on the copyright holders.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 #include <unistd.h>
 #include <sys/socket.h>
 #include <sys/types.h>
@@ -6,7 +23,8 @@
 
 #include "ThreadPool.H"
 #include "SocketInput.H"
-
+#include "InputChannel.H"
+#include "Exception.H"
 
 
 // create unix domain socket for input
@@ -21,6 +39,7 @@ SocketInput::SocketInput(const char *path,
 	saddr.sun_family = AF_UNIX;
 	strcpy(saddr.sun_path, path);
 	fd = socket(PF_UNIX, SOCK_STREAM, 0);
+	if(fd < 0) throw new Exception;
 	if(connect(fd, (struct sockaddr*)&saddr, sizeof(saddr.sun_path)) < 0) {
 		if(errno == ECONNREFUSED) {
 			unlink(saddr.sun_path);
@@ -29,9 +48,10 @@ SocketInput::SocketInput(const char *path,
 		// another instance running
 		// throw new Exception
 	}
-	bind(fd, (struct sockaddr *)&saddr, sizeof(saddr));
-	listen(fd, SOCK_QUEUE_MAX);
-	ThreadPool::instance()->setWorkAccept(this);
+	if(bind(fd, (struct sockaddr *)&saddr, sizeof(saddr)) < 0) 
+		throw new Exception;
+	if(listen(fd, SOCK_QUEUE_MAX) < 0)
+		throw new Exception;
 }
 
 
@@ -48,8 +68,9 @@ void
 SocketInput::onReady()
 {
 	Connection *conn = cFactory->accept(fd);
-	Transport  *trans = tFactory->newTransport(conn);
-	ThreadPool::instance()->queueWorkRead(trans);
+	Transport  *trans = tFactory->newTransport();
+	InputChannel *channel = new InputChannel(conn, trans);
+	channel->start();
 }
